Make shredding priv_key idempotent

+++ infrastructure/scripts/login-gh.sh
@@ -6,8 +6,10 @@ echo " === login-gh.sh === "
 
 if [[ "$GIT_AUTH_VIA_SSH_KEY" != true ]]; then
   # remove private keys since user opted into GitHub Auth
-  shred ~/.ssh/priv_key
-  rm ~/.ssh/priv_key
+  if [[ -f ~/.ssh/priv_key ]]; then
+    shred ~/.ssh/priv_key
+    rm ~/.ssh/priv_key
+  fi
 
   # use GitHub token for git auth via the environment
   echo "export GH_USERNAME=\"$GH_USERNAME\"" >> .envrc
